Moved indexing ops to indexing.h
