+++ csrc/ops/all_ops.h
@@ -9,4 +9,5 @@
 #include <ops/alias.h>
 #include <ops/arith.h>
 #include <ops/composite.h>
+#include <ops/indexing.h>
 #include <ops/normalization.h>
