+++ csrc/ops/arith.cpp
@@ -110,189 +110,6 @@ TensorView* unaryOp(
   return unaryOp(type, cast_v1)->as<TensorView>();
 }
 
-TensorView* select(TensorView* tv, int dim, Val* index) {
-  auto dom = TensorDomain::noReductions(tv->getMaybeRFactorDomain());
-  TORCH_CHECK(!dom.empty(), "select can not be applied to 0d tensor.");
-
-  std::vector<IterDomain*> new_root;
-  new_root.reserve(dom.size() - 1);
-
-  if (dim < 0) {
-    dim += (int)dom.size();
-  }
-
-  TORCH_CHECK(
-      dim >= 0 && dim < (int)dom.size(),
-      "Select on invalid axis, received: ",
-      dim,
-      " however tensor view only has ",
-      dom.size(),
-      " non-reduction dims.");
-
-  for (auto i : c10::irange(dom.size())) {
-    if ((int)i != dim) {
-      new_root.emplace_back(dom[i]->cloneWithoutRFactor());
-    }
-  }
-
-  auto td = IrBuilder::create<TensorDomain>(
-      new_root, TensorDomain::getContiguityFilledWith(new_root, true));
-  auto out = IrBuilder::create<TensorView>(td, *tv->getDataType());
-  IrBuilder::create<SelectOp>(out, tv, dom.at(dim), index);
-  return out;
-}
-
-// index_select
-TensorView* index_select(TensorView* lookup_tv, int dim, TensorView* index_tv) {
-  DataType dtype = lookup_tv->getDataType().value();
-  TORCH_CHECK(
-      dtype != DataType::Null, "Invalid datatype provided for new value.");
-  auto lookup_dom =
-      TensorDomain::noReductions(lookup_tv->getMaybeRFactorDomain());
-  auto index_dom =
-      TensorDomain::noReductions(index_tv->getMaybeRFactorDomain());
-  size_t n_dims = lookup_dom.size();
-  TORCH_CHECK(n_dims > 0, "index_select can not be applied to 0d tensor.");
-  TORCH_CHECK(
-      index_dom.size() <= 1, "index array must be 1d or scalar tensor.");
-
-  if (index_dom.empty()) {
-    auto select_tv = select(lookup_tv, dim, index_tv);
-    return unsqueeze(select_tv, dim);
-  }
-
-  if (dim < 0) {
-    dim += (int)lookup_dom.size();
-  }
-
-  std::vector<IterDomain*> new_root;
-  new_root.reserve(lookup_dom.size() - 1);
-  TORCH_CHECK(
-      dim >= 0 && dim < (int)lookup_dom.size(),
-      "index_select on invalid axis, received: ",
-      dim,
-      " however tensor view only has ",
-      lookup_dom.size(),
-      " non-reduction dims.");
-
-  for (auto i : c10::irange(lookup_dom.size())) {
-    if ((int)i != dim) {
-      new_root.emplace_back(lookup_dom[i]->cloneWithoutRFactor());
-    } else {
-      new_root.emplace_back(index_dom[0]->cloneWithoutRFactor());
-    }
-  }
-
-  auto td = IrBuilder::create<TensorDomain>(
-      new_root, TensorDomain::getContiguityFilledWith(new_root, true));
-  auto out = IrBuilder::create<TensorView>(td, dtype);
-
-  // broadcast index to lookup's rank.
-  index_tv =
-      ops::maybe_broadcast_index_tv(index_tv->as<TensorView>(), dim, n_dims);
-  IrBuilder::create<IndexSelectOp>(
-      out, lookup_tv, dim, lookup_dom[dim], index_tv);
-  return out;
-}
-
-// torch.gather
-TensorView* torch_gather(TensorView* inp, int dim, TensorView* index) {
-  auto inp_domain = TensorDomain::noReductions(inp->getMaybeRFactorDomain());
-  auto idx_domain = TensorDomain::noReductions(index->getMaybeRFactorDomain());
-  TORCH_CHECK(
-      !inp_domain.empty(), "torch.gather can not be applied to 0d tensor.");
-  TORCH_CHECK(
-      idx_domain.size() == inp_domain.size(),
-      "the input and index tensor must have the same dimensions for torch.gather");
-
-  if (dim < 0) {
-    dim += (int)idx_domain.size();
-  }
-  TORCH_CHECK(
-      dim >= 0 && dim < (int)inp_domain.size(),
-      "torch.gather on invalid axis, received: ",
-      dim,
-      " however tensor view only has ",
-      inp_domain.size(),
-      " non-reduction dims.");
-  std::vector<IterDomain*> out_domain;
-  out_domain.reserve(idx_domain.size());
-  for (auto idx_domain_ptr : idx_domain) {
-    out_domain.push_back(
-        IterDomainBuilder(idx_domain_ptr)
-            .iter_type(
-                idx_domain_ptr->getIterType() == IterType::Iteration
-                    ? IterType::GatherScatter
-                    : idx_domain_ptr->getIterType())
-            .build());
-  }
-
-  TensorView* out_tensor = IrBuilder::create<TensorView>(
-      IrBuilder::create<TensorDomain>(
-          out_domain, TensorDomain::getContiguityFilledWith(out_domain, true)),
-      inp->getDataType().value());
-
-  IrBuilder::create<TorchGatherOp>(
-      out_tensor, inp, dim, inp_domain[dim], index);
-  return out_tensor->as<TensorView>();
-}
-
-// torch.scatter torch.scatter_add
-TensorView* scatterOp(
-    ScatterOpType type,
-    TensorView* self,
-    int dim,
-    TensorView* index,
-    TensorView* src) {
-  auto self_dom = TensorDomain::noReductions(self->getMaybeRFactorDomain());
-  auto idx_dom = TensorDomain::noReductions(index->getMaybeRFactorDomain());
-  auto src_dom = TensorDomain::noReductions(src->getMaybeRFactorDomain());
-
-  TORCH_CHECK(!self_dom.empty(), "scatter can not be applied to 0d tensor.");
-  TORCH_CHECK(
-      self_dom.size() == idx_dom.size() && self_dom.size() == src_dom.size(),
-      "self, index and src tensor should all have the same number of dimensions in scatter like ops.");
-  if (dim < 0) {
-    dim += (int)self_dom.size();
-  }
-  TORCH_CHECK(
-      dim >= 0 && dim < (int)self_dom.size(),
-      "Scatter on invalid axis, received: ",
-      dim,
-      " however tensor view only has ",
-      self_dom.size(),
-      " non-reduction dims.");
-
-  // The shape of output tensor is same as self tensor.
-  std::vector<IterDomain*> out_domain;
-  for (const auto i : c10::irange(self_dom.size())) {
-    out_domain.push_back(
-        IterDomainBuilder(self_dom[i])
-            .iter_type(
-                self_dom[i]->getIterType() == IterType::Iteration
-                    ? IterType::GatherScatter
-                    : self_dom[i]->getIterType())
-            .build());
-  }
-
-  TensorView* out_tensor = IrBuilder::create<TensorView>(
-      IrBuilder::create<TensorDomain>(
-          out_domain, TensorDomain::getContiguityFilledWith(out_domain, true)),
-      self->getDataType().value());
-
-  IrBuilder::create<ScatterOp>(
-      type, out_tensor, self, dim, index, src, out_domain[dim]);
-  return out_tensor->as<TensorView>();
-}
-
-TensorView* scatter(
-    TensorView* self,
-    int dim,
-    TensorView* index,
-    TensorView* src) {
-  return scatterOp(ScatterOpType::Set, self, dim, index, src);
-}
-
 // TENSOR FACTORIES
 TensorView* rand(const std::vector<Val*>& shape, DataType dtype) {
   auto n = shape.size();
