+++ csrc/python_frontend/fusion_record.h
@@ -8,9 +8,7 @@
 #pragma once
 #include <c10/util/complex.h>
 #include <ir_interface_nodes.h>
-#include <ops/alias.h>
-#include <ops/arith.h>
-#include <ops/normalization.h>
+#include <ops/all_ops.h>
 #include <python_frontend/fusion_definition.h>
 #include <python_frontend/fusion_state.h>
 #include <serde/fusion_record_serde.h>
